requirements.txt
psycopg2==2.8.6
Flask==1.1.2
PyJWT==2.0.0
